+++ packages/espressocash_api/lib/src/dto/durable_transactions.dart
@@ -1,7 +1,7 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-part 'durable_transactions.g.dart';
 part 'durable_transactions.freezed.dart';
+part 'durable_transactions.g.dart';
 
 @freezed
 class GetFreeNonceResponseDto with _$GetFreeNonceResponseDto {
@@ -33,4 +33,16 @@ class SubmitDurableTxResponseDto with _$SubmitDurableTxResponseDto {
 
   factory SubmitDurableTxResponseDto.fromJson(Map<String, dynamic> json) =>
       _$SubmitDurableTxResponseDtoFromJson(json);
+}
+
+@freezed
+class GetDurableFeesResponseDto with _$GetDurableFeesResponseDto {
+  const factory GetDurableFeesResponseDto({
+    required int outgoingLink,
+    required int incomingLink,
+    required int cancelLink,
+  }) = _GetDurableFeesResponseDto;
+
+  factory GetDurableFeesResponseDto.fromJson(Map<String, dynamic> json) =>
+      _$GetDurableFeesResponseDtoFromJson(json);
 }