+++ packages/espressocash_app/lib/features/authenticated/widgets/balance_amount.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import '../../../di.dart';
 import '../../../l10n/device_locale.dart';
 import '../../../ui/value_stream_builder.dart';
-import '../../conversion_rates/services/watch_token_fiat_balance.dart';
+import '../../conversion_rates/services/token_fiat_balance_service.dart';
 import '../../conversion_rates/widgets/extensions.dart';
 import '../../currency/models/amount.dart';
 import '../../currency/models/currency.dart';
@@ -17,9 +17,7 @@ class BalanceAmount extends StatelessWidget {
   @override
   Widget build(BuildContext context) => ValueStreamBuilder<Amount>(
         create: () => (
-          sl<WatchTokenFiatBalance>()
-              .call(Token.usdc)
-              .map((event) => event ?? Amount.zero(currency: Currency.usd)),
+          sl<TokenFiatBalanceService>().watchMainBalance(),
           Amount.zero(currency: Currency.usd),
         ),
         builder: (context, amount) {