+++ packages/espressocash_app/lib/features/incoming_link_payments/data/ilp_repository.dart
@@ -92,18 +92,16 @@ extension on ILPStatusDto {
   ILPStatus toModel(ILPRow row) {
     final tx = row.tx?.let(SignedTx.decode);
     final txId = row.txId;
-    final slot = row.slot?.let(BigInt.tryParse);
 
     switch (this) {
       case ILPStatusDto.txCreated:
         return ILPStatus.txCreated(
           tx!,
-          slot: slot ?? BigInt.zero,
         );
       case ILPStatusDto.txSent:
         return ILPStatus.txSent(
           tx ?? StubSignedTx(txId!),
-          slot: slot ?? BigInt.zero,
+          signature: row.txId!,
         );
       case ILPStatusDto.success:
         final feeAmount = row.feeAmount;
@@ -130,7 +128,6 @@ extension on IncomingLinkPayment {
         status: status.toDto(),
         tx: status.toTx(),
         txId: status.toTxId(),
-        slot: status.toSlot()?.toString(),
         txFailureReason: status.toTxFailureReason(),
         feeAmount: switch (status) {
           ILPStatusSuccess(:final fee) => fee?.value,
@@ -153,15 +150,11 @@ extension on ILPStatus {
       );
 
   String? toTxId() => mapOrNull(
+        txSent: (it) => it.signature,
         success: (it) => it.tx.id,
       );
 
   TxFailureReason? toTxFailureReason() => mapOrNull(
         txFailure: (it) => it.reason,
-      );
-
-  BigInt? toSlot() => mapOrNull(
-        txCreated: (it) => it.slot,
-        txSent: (it) => it.slot,
       );
 }