+++ packages/espressocash_app/CHANGELOG.md
@@ -1,3 +1,30 @@
+## 1.122.1
+
+ - **FIX**: durable tx signing.
+
+## 1.122.0
+
+ - **FIX**: home spacing (#1489).
+ - **FEAT**: use durable nonce for escrow txs (#1459).
+ - **FEAT**: integrate stellar wallet (#1497).
+ - **FEAT**: new lock screen theme  (#1494).
+
+## 1.121.1
+
+ - **FIX**: do not wait for intercom login (#1493).
+ - **FIX**: profile spacing colors (#1490).
+ - **FIX**: sol conversion matching (#1491).
+
+## 1.121.0
+
+ - **REFACTOR**: balances module (#1476).
+ - **FIX**: main screen flow (#1487).
+ - **FIX**: ignore usdc token on crypto portfolio (#1478).
+ - **FIX**: activity tile text color to white (#1477).
+ - **FEAT**: log total crypto balance (#1488).
+ - **FEAT**: sort crypto portfolio list (#1480).
+ - **FEAT**: home & payment improved buttons (#1468).
+
 ## 1.120.0
 
  - **REFACTOR**: request feature flag cleanup (#1455).